+++ linera-execution/src/system.rs
@@ -15,7 +15,7 @@ use linera_base::{
     crypto::{CryptoHash, PublicKey},
     data_types::{Amount, ArithmeticError, Timestamp},
     ensure, hex_debug,
-    identifiers::{BytecodeId, ChainDescription, ChainId, MessageId, Owner},
+    identifiers::{Account, BytecodeId, ChainDescription, ChainId, MessageId, Owner},
 };
 
 #[cfg(with_metrics)]
@@ -35,7 +35,6 @@ use std::{
     collections::BTreeMap,
     fmt::{self, Display, Formatter},
     iter,
-    str::FromStr,
 };
 use thiserror::Error;
 
@@ -353,59 +352,6 @@ impl Recipient {
     #[cfg(any(test, feature = "test"))]
     pub fn root(index: u32) -> Recipient {
         Recipient::chain(ChainId::root(index))
-    }
-}
-
-/// A system account.
-#[derive(Debug, PartialEq, Eq, Hash, Copy, Clone, Serialize, Deserialize)]
-pub struct Account {
-    /// The chain of the account.
-    pub chain_id: ChainId,
-    /// The owner of the account, or `None` for the chain balance.
-    pub owner: Option<Owner>,
-}
-
-impl Account {
-    pub fn chain(chain_id: ChainId) -> Self {
-        Account {
-            chain_id,
-            owner: None,
-        }
-    }
-
-    pub fn owner(chain_id: ChainId, owner: Owner) -> Self {
-        Account {
-            chain_id,
-            owner: Some(owner),
-        }
-    }
-}
-
-impl std::fmt::Display for Account {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self.owner {
-            Some(owner) => write!(f, "{}:{}", self.chain_id, owner),
-            None => write!(f, "{}", self.chain_id),
-        }
-    }
-}
-
-impl FromStr for Account {
-    type Err = anyhow::Error;
-
-    fn from_str(s: &str) -> Result<Self, Self::Err> {
-        let parts = s.split(':').collect::<Vec<_>>();
-        anyhow::ensure!(
-            parts.len() <= 2,
-            "Expecting format `chain-id:address` or `chain-id`"
-        );
-        if parts.len() == 1 {
-            Ok(Account::chain(s.parse()?))
-        } else {
-            let chain_id = parts[0].parse()?;
-            let owner = parts[1].parse()?;
-            Ok(Account::owner(chain_id, owner))
-        }
     }
 }
