+++ linera-service/src/linera/client_options.rs
@@ -7,12 +7,11 @@ use chrono::{DateTime, Utc};
 use linera_base::{
     crypto::PublicKey,
     data_types::Amount,
-    identifiers::{BytecodeId, ChainId, MessageId},
+    identifiers::{Account, BytecodeId, ChainId, MessageId},
 };
 use linera_execution::{
-    committee::ValidatorName,
-    system::{Account, SystemChannel},
-    UserApplicationId, WasmRuntime, WithWasmDefault,
+    committee::ValidatorName, system::SystemChannel, UserApplicationId, WasmRuntime,
+    WithWasmDefault,
 };
 use linera_service::{
     chain_listener::{ChainListenerConfig, ClientContext as _},
